+++ src/mscorlib/shared/System/Runtime/CompilerServices/ValueTaskAwaiter.cs
@@ -4,6 +4,7 @@
 
 using System.Diagnostics;
 using System.Threading.Tasks;
+using System.Threading.Tasks.Sources;
 
 namespace System.Runtime.CompilerServices
 {
