+++ src/mscorlib/shared/System/Threading/Tasks/ValueTask.cs
@@ -6,6 +6,8 @@ using System.Collections.Generic;
 using System.Diagnostics;
 using System.Runtime.CompilerServices;
 using System.Runtime.InteropServices;
+using System.Threading.Tasks.Sources;
+
 #if !netstandard
 using Internal.Runtime.CompilerServices;
 #endif
